+++ libs/components/datetime/src/lib/modules/datepicker/datepicker-input-fuzzy.directive.ts
@@ -347,8 +347,8 @@ export class SkyFuzzyDatepickerInputDirective
     }
   }
 
-  @HostListener('keyup')
-  public onInputKeyup(): void {
+  @HostListener('input')
+  public onInput(): void {
     this.control.markAsDirty();
   }