+++ libs/components/forms/testing/src/public-api.ts
@@ -1,4 +1,9 @@
 export * from './input-box/input-box-harness-filters';
 export * from './input-box/input-box-harness';
+
 export * from './checkbox-fixture';
+export * from './checkbox/checkbox-harness';
+export * from './checkbox/checkbox-harness-filters';
+export * from './checkbox/checkbox-label-harness';
+
 export * from './radio-fixture';