+++ libs/components/layout/src/lib/modules/format/format-item.ts
@@ -6,5 +6,5 @@ import { TemplateRef } from '@angular/core';
 export interface SkyFormatItem {
   text?: string;
 
-  templateRef?: TemplateRef<unknown>;
+  templateRef?: TemplateRef<any>;
 }